mlflow_model_serve/requirements.txt
alembic==1.4.1
anyio==3.4.0
asgiref==3.4.1
boto3==1.20.24
botocore==1.23.24
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
databricks-cli==0.16.2
docker==5.0.3
entrypoints==0.3
fastapi==0.70.1
Flask==2.0.2
fonttools==4.28.3
gitdb==4.0.9
GitPython==3.1.24
gunicorn==20.1.0
h11==0.12.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
kiwisolver==1.3.2
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mlflow==1.22.0
numpy==1.21.4
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.6
protobuf==3.19.1
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
querystring-parser==1.2.4
requests==2.26.0
s3transfer==0.5.0
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.4.28
sqlparse==0.4.2
starlette==0.16.0
tabulate==0.8.9
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websocket-client==1.2.3
websockets==10.1
Werkzeug==2.0.2
zipp==3.6.0
